Docker-Compose-Based S3 for Local Development (#63019)
* Adds basic docker-compose.yml for development environment.

* Adds DOCKER.md.

* Adds docker/developers path to the dependabot metadata.

* Update DOCKER.md

Co-authored-by: Stephen Liang <stephenliang@users.noreply.github.com>

* Allows for port to be specified in the database host.

This adds the ability to specify a port via typical ':' syntax to the db_endpoint_* configuration fields. This will allow you to more easily spin up secondary databases for secondary development environments. In particular, this enhances the Docker version of things my allowing the port to be specified via `MYSQL_NATIVE_PORT` such that you can spin it up alongside a native database more easily (and defer uninstalling one if you are migrating.)

This requies some small changes in our utility scripts for getting a console-based REPL for the database connections.

This also updates references to the database user in the Docker context to use `root` since that was what was used to create the databases.

* Makes the default mysql port 3306.

* Adds redis port configuration for natively run redis.

* Adds docker/developers path to the dependabot metadata.

* move docker-specific documentation into docker-specific directory

* add links to official documentation, extract some steps common to each environment, and slightly reorganize

* move daily reference details above first-time setup details

* add note to stop native mysql and redis services if they exist

* misc markdown formatting tweaks

more misc markdown formatting tweaks

* Adds MinIO support and adds guards for S3 emulation.

Enables Dance Party and most labs without AWS access.

* Guard cdo-ai validation to only seed runs that have AWS access.

* Removes print statements.

* Adds info to DOCKER.md.

* simplify install-s3.sh

* move logic for populating emulated s3 out of docker directory

* remove local file cache; persist downloaded files only to local S3

* comments & cleanup

* misc cleanup

* restore access key config entries, but scope the naming

* harden fragile env check

* Update README.md

Fix broken image link

* update setup instructions to reflect renamed config values

* Add environment-unaware `aws_s3_emulated` config option

To replace the environment-aware helper method

* apply some misc documentation/comment improvements

---------

Co-authored-by: wilkie <david.wilkinson@code.org>
Co-authored-by: Stephen Liang <stephenliang@users.noreply.github.com>