Give Replication Groups Stack-Specific Descriptions (#63712)
We for the most part only use these components in the production
environment, but every so often we do spin up a full-stack adhoc and it
would in general be nice to have better support for more prod-like
environments. To that end, update the description for each of them to
add stack-specific details.

Note that [the `ReplicationGroupDescription` property can be updated
with no
interruption](https://docs.aws.amazon.com/AWSCloudFormation/latest/UserGuide/aws-resource-elasticache-replicationgroup.html#cfn-elasticache-replicationgroup-replicationgroupdescription)

It would probably be preferable to give these entities more specific
names than the current randomly-generated one, but it's not clear to me
how to do that without service interruption.