feat(source-maps): add source maps to component library
This PR adds source maps to the component library. For `marketing`,
source maps are automatically loaded by Next.js. For `apps`, source maps
need to be loaded via the `source-map-loader`. While it might make sense
to enable source maps for all node modules in development mode, `apps`
is already fairly slow so this PR limits the scope of the source maps to
just packages in `frontend/packages`.