feat(marketing): add concurrency limit
This PR limits the concurrency of github actions for the marketing app.
Currently, when multiple commits are pushed to `staging`, the Github
Action is executed in parallel which results in an update to the stack
while an update is already occurring. Additionally, it results in two
docker builds which don't contain the other commit.

To mitigate this, pushes to staging are queued up while pull requests
continue to run in parallel.