fix(jest): limit max workers to 75% of cpu cores
The Code.org CI environment is a shared machine which is subject to
resource contention. When cron jobs, asynchronous jobs, and other tasks
run while a jest suite executes, there is a possibility one of the cores
are running other resource intensive tasks resulting in a lengthy delay
before the core is returned to the worker, thus causing it to exceed the
5s timeout. Since the CI environment is shared, limit the maximum
workers to 75% of cores to allow some to remain to be used by other
processes.