Improve "Unknown property not in defaults" Error Message (#63931)
The bespoke error message we surface when an environment is using a
configuration property not included in `config.yml.erb` is somewhat
generic. Update it to be a bit more specific about the nature of the
problem, and to provide a recommendation for resolution.

Also tweak it to handle multiple violations simultaneously.