feat(stylelint): add stylelint support
This PR copies the stylelint configuration from `apps` and is now used
as the global lint configuration for scss/sass/css. A follow up will be
to remove the stylelint configuration from `apps` and use the one in the
linked lint config.

Many rules were already disabled in the stylelint configuration that
existed in `apps`. A follow up improvement is to replace the disabled
rules with disabled lines so that new code always follows the
recommended rules. That being said, this will likely require further
discussions prior to changing the rule and as such, this PR does not
change the rule set.