chore: reduce jest workers to 50%
Staging continues to have jest build failures likely due to resources
running out. We previously attempted to remediate this by decreasing
workers to 75% in #63955 and it does seem to reduce the number of failing
tests, but not to a stable level. This PR further reduces workers to
50%, which is likely the lowest amount before tests start to take really
long. If this still does not work, some deeper analysis would be
required to understand why so much resources is being consumed.