DTL (Test > Levelbuilder: a191e819) (#64104)
* feat(frontend): add precommit linting support

This PR adds the ability for the `frontend` directory to be prettier
fixed and linter fixed prior to committing. Since `frontend` is a
monorepo, each app and package's linting is delegated to its own linter
and prettier configuration which means the ruby linter strategy for `apps`
cannot be applied to the `frontend` directory.

Rather, this PR leverages a hierarchical approach to linting. When a
staged file is set to be committed, the nearest `.lintstagedrc.mjs` will
be read triggering the nearest eslint and prettier configurations to be
read. This hierarchical implementation is important as different
packages have different rule sets. For example, Storybook will want
react based linting whereas `component-styles` needs style linting and
no eslint.

* Blockly: update dark trash color

* fix(Video): add ignore region for YT logo

* [Pythonlab] Add basic backpack feature under experiment  (#64035)

[Pythonlab] Add basic backpack feature under experiment

* changed cronjob frequency and added checkout (#63867)

* staging content changes (-robo-commit)

* Revert "Revert "[CMS-358] DSCO - Revert revert [Design2-263] DSCO - Typograph…"

This reverts commit e02a9b0b5b713d550e5156b4e86d7efffa128128.

* fix(font.scss): fixed barlowFont variables definitions

* feat(stylelint): add stylelint support

This PR copies the stylelint configuration from `apps` and is now used
as the global lint configuration for scss/sass/css. A follow up will be
to remove the stylelint configuration from `apps` and use the one in the
linked lint config.

Many rules were already disabled in the stylelint configuration that
existed in `apps`. A follow up improvement is to replace the disabled
rules with disabled lines so that new code always follows the
recommended rules. That being said, this will likely require further
discussions prior to changing the rule and as such, this PR does not
change the rule set.

* feat(paragraph): added color options

* fix(Carousel): wait for carousel load to complete

The original implementation of the test did not wait for the promise to
complete, thus if there's a timing issue where the component is still
loading, the test may fail. Instead, `await` the promise.

* feat(workshops): replace virtual attribute with method (#63697)

* feat: remove 'virtual' workshop attribute and replace it with a function that checks if any of the workshop sessions are virtual

* chore: remove friday institute specific logic

* chore: remove tests around setting workshop.virtual field which can no longer be set or updated

* fix lint failure

* chore: remove monoweave

* feat(dsco): add support for conventional commits

This PR adds the ability to use conventional commits to automatically
generate changelogs and bump versions using the `release-it` library.

When the maintainer wants to publish a new version of the package,
execute the `yarn release` script which will then scan from the last
tagged version of the package in git followed by scanning conventional
commits. Those are then populated into `CHANGELOG.md` and the package
version bumped.

* Getting a new MusicLab Instructions arrow for better visibility (#63971)

* new arrow

* second image

* new arrow png

* adding back in teal arrow for music toolbox

* add AI Tutor Tab to teacher nav (#63896)

* add AI Tutor Tab to teacher nav

* bug fix

* move ai tutor redirect

* revert v1 teacher dashboard

* update redirect logic

* add 'programming-fundamentals-aitutor-2024' to AI tutor tab access list

* added DCDO flag

* bug fix

* Improvements to file tabs and add tests (#64069)

* work in progress

* working

* activate file on drag start

* work in progress

* fix bugginess

* maybe an improvement

* remove console log

* add close test

* increase timeout

* remove role and update test

* Disallow indexing of non production environments on studio.code.org (#63966)

* Add projects route to disallow list in robots.txt

* reorder list

* Disallow all on non production environments in robots.txt

* Refactor way we do robots.txt to use controller

* Clean up file

* Remove unwanted file changes

* Fix tests

* Remove unwanted changes

* Fix tests more

* Add pr suggestions

* update dcdo on backend (#64098)

* Remove experimental suffix from gpt-4o-mini model name (#64061)

* AI Chat: use OpenAI's full 0-2 temperature scale, update existing levels to match (#64060)

* Scale gpt-4o-mini temperature to 0-2

* Script to update AI Chat temperatures for levels using gpt-4o-mini

* Puts on save

* Level changes from running script cutting temperature in half

* Revert "Level changes from running script cutting temperature in half"

This reverts commit a17b4eca30d59df0d933cfc6901807759ee236ff.

* Cut temperature in half again, including newly created levels

---------

Co-authored-by: Stephen Liang <538214+stephenliang@users.noreply.github.com>
Co-authored-by: breville <brendan@code.org>
Co-authored-by: fisher-alice <107423305+fisher-alice@users.noreply.github.com>
Co-authored-by: Stephen Liang <stephenliang@users.noreply.github.com>
Co-authored-by: Code.org Deploy Systems <dev@code.org>
Co-authored-by: Mario Gil Correa <66776217+mgc1194@users.noreply.github.com>
Co-authored-by: levadadenys <levada.denys@gmail.com>
Co-authored-by: Ryan Driscoll <ryan.driscoll@code.org>
Co-authored-by: Hannah Bergam <hannahbergam@gmail.com>
Co-authored-by: Mark Barnes <mark.barnes@code.org>
Co-authored-by: Molly Moen <molly@code.org>
Co-authored-by: Dani LaMarca <dani@code.org>
Co-authored-by: Brendan Reville <breville@users.noreply.github.com>