fix(fonts): remove multiple declarations of barlow
Barlow is defined in the SCSS Module for Typography, which means the
font face is redefined each time typography is used. This results in a
font-face declaration in _each_ usage of Typography. To remove the
duplicate declarations, remove the font-face declaration from
Typography.

Additionally, Barlow is already defined as a global font in
[application.scss](https://github.com/code-dot-org/code-dot-org/blob/d207451b6260de67d03eb87d531ae5d0fa43e8ef/dashboard/app/assets/stylesheets/application.scss#L24-L27)