feat(build): use native typescript compiler
The default for esbuild/tsup is to use the microsoft api extractor,
to generate typescript declarations. However, this appears to result in
some memory leaks which ends up causing the build time of declaration
files to take very long. This then causes watchers on the webpack side
to receive updated javascript but without its associated declarations.

Instead, noting that the [official documentation](https://tsup.egoist.dev/#generate-declaration-file) says
to use `tsc`, the tsup `onSuccess` hook is then used to generate the
typescript declarations in the same manner as before, except using
`tsc`. The net result is that the declarations are generated
significantly faster, about 2 seconds vs the previous 30-60 seconds.