fix(packaging): also hash `frontend` when packaging apps
The CI/CD process includes a git hash mechanism to detect if a
precompiled version of `apps` is available on S3. If so, it will
download that compiled version and extract it locally skipping a local
build. This mechanism uses a git file hashes for `apps`. However, now
that we have a separate repo called `frontend`, changes to this
directory also need to be reflected in the calculated hash otherwise the
following race condition can occur:

**Stage 1:**
1. Push a change specific to `apps`, no changes to `frontend`.
2. A hash is generated based on the hashes of the files in `apps`.
2. Create a build for `apps`, triggering its bundler process.
3. The build output is zipped and stored as the hash generated in (2).

**Stage 2:**
1. Push a change specific to `frontend`, no changes to `apps`.
2. A hash is generated based on the hashes of the files in `apps`.
3. No build is performed as the hash will match one on S3 from Stage 1.

There are a couple variations to this race condition but they all
result in a build that does not incorporate changes from `frontend` if
the changes occurred outside of `apps` only (cross changes will trigger
a build).

To fix this, the hashing process should include all files that
contribute to `apps`.