feat(lighthouse): add lighthouse to ci process
This PR adds support for Google Lighthouse[1] checks during our CI/CD
process. Google Lighthouse adds performance, SEO, accessibility, and
other checks that are important for the marketing site. The marketing
site aims to be WCAG AA compliant and search engine optimized and
therefore this check aims to maintain scores at a baseline.

Currently, the marketing site receives a fairly poor score in
performance but we plan to address performance in a subsequent follow
up. Rather, this PR seeks to prevent the scores from further slipping by
establishing the baselines that exist as of today. We will seek to
improve them over time and continue to raise baseline scores.

[1] https://developer.chrome.com/docs/lighthouse/overview