feat(eyes): implement eyes batching for `frontend`
This PR adds the ability to consolidate distributed eyes tests into the
same batch within eyes. This is needed because Applitools reports
success/fail status on a single batch to Github using its commit id.
Previously, eyes tests were executing in their own batches - `marketing`
and `component library` which eyes will then report whichever reported
first.

Rather, the workflow has been refactored to utilize a common entrypoint,
the `frontend-ci` workflow, which will execute on pull requests. The
previous workflows are no longer children of this main workflow and will not
be triggered when pull requests are executed. This allows a common
`setup` and `teardown` job to execute _before_ the execution of the
workflow.

The `setup` workflow will establish common environmental or other setup
tasks necessary to enable multiple workflows to execute. The `teardown`
job will then perform clean up actions. In the context of eyes batching,
it will close the batch notating its completion. This then triggers eyes
to report success/fail status in the "checks" section of Github.

As a side effect, this produces a good visualization of all the steps
that are executed.