feat(Video): replace with react-player
Previously, the Video component directly used the YouTube iframe embed
API to load the YouTube video page. This is a direct port of the
existing Video loading mechanism in Pegasus. However, for pages that
load many YouTube videos (such as /videos), the page loads very slowly
due to the iframe necessitating the loading of the bundle for each
frame.

Rather, following the Google Lighthouse best practice for [loading third-party resources](https://developer.chrome.com/docs/lighthouse/performance/third-party-facades),
a new library has been added, `react-player`, which enables a facade of
the video to be loaded. What this means, in effect, is that rather than
loading frames of YouTube multiple times, the Video component will
render out an image and a play button. Only when the user wishes to
click on a specific video will the facade be substituted with the actual
frame.

This solves a number of current issues we face for the Video component:

1. The YouTube logo frequently swaps between bottom left and bottom
right resulting in test flakiness. This is fixed by using our own facade.
2. Pages with videos load very slowly due to network congestion caused
by loading each YouTube frame optimistically.
3. When YouTube is blocked, whether due to school or geographic
firewalls, or a computer/browser setting we do not gracefully fallback
to the native video player in all circumstances. In this new
implementation, the video player will fallback to the native player if
any error (except `NotAllowedError`, which the user can self-service by
clicking on the video).