fix(storybook): add build target
The component library was failing to deploy because the `test:ui:ci`
target was being cached in previous runs. To address this, do not cache
a UI test which requires revalidation due to its dependencies
potentialyl changing.

Additionally, the deployment also failed because the deployment requires
the `dist` folder to be created. The `build` target is supposed to
generate this folder but it wasn't defined for storybook. Instead,
`test:ui:ci` generated the folder by way of the `build-storybook`
target. Since the `test:ui:ci` target was cached, it never generates the
`dist` folder.