feat(OneTrust): add OneTrust support to `marketing`.
This PR integrates the marketing app with OneTrust, the Code.org cookie
management platform. As a general note of how OneTrust works, when OneTrust
detects an external `src`, it overwrites the tag with `text/plain`, which
effectively causes the tag to not make a 3P request. When consent is
provided for a specific cookie, OneTrust will restore the appropriate
tag. When programming this, it's important to note that OneTrust may or
may not exist as it can be blocked so the types reflect that it can be
`undefined`.

This PR adds a couple helper functions to make the usage of OneTrust
easier within Code.org:

1. A Context Provider has been added which can read the client-side OneTrust settings.
2. Client-side components can determine whether a feature is enabled using the context provider.
3. When consent changes, the context provider will update appropriately.

Since YouTube injects 3P cookies, this PR also refactors the component
to not load YouTube unless the user has explicitly provided consent for
"Functional Cookies".