feat(marketing): upload static assets to s3
This PR adds the ability to upload the static assets within the
nextjs marketing build output to S3 and enables switching over those
assets from being served from the application to being served by a CDN.
This PR does not implement the Cloudfront connection to S3, but rather
it just adds the ability to upload the assets.

To accomplish multi-environment deployments, this PR uses the built
docker image that is produced in the build step and during a deployment,
extracts the static assets out from the image - effectively making the
docker image the source of truth and enabling consistency when running
on the Docker-only version and the cloud-optimized version on AWS.

To prevent this bucket from becoming a dumping ground, old assets will
automatically be deleted after 90 days. Since assets are uploaded using
`aws s3 cp`, even if an asset never changes, it will create a new
version of that asset. After 90 days, the old versions of the asset are
deleted as they are no longer current.