feat(efs): add distributed EFS for full page cache
This PR supplements the ISR PR by adding distributed volume mounting via
Amazon Elastic Filesystem (EFS). Next.js caches full page router paths
via the `.next/server/app/[brand]` path in the form of local file
storage. To achieve consistent caching throughout the cluster, EFS was
added such that one cache update will distribute to all instances.

This is needed because the existing strategy of invalidating cache
entries during a deployment no longer applies. Caching invalidation
needs to occur when an entry, asset, or page is published on Contentful
which necessitates the use of a page revalidation strategy. This allows
for pages to be revalidated via API calls but without a distributed
cache, making an API call to invalidate a particular path will only
invalidate the cache on **one** instance and the other instances will
not be invalidated.

Using EFS solves this problem by causing the cache update to surface
across all instances in a consistent manner.

Additionally, this pattern was adopted using the AWS Labs CDK Next.js
implementation but ported over to raw Cloudformation. This helps to
ensure our implementation of Next.js on AWS is following the same best
practice princinples that are being advocated by AWS.