chore(marketing/deploy): enable multi-stack deployments
This PR refactors the marketing stack to allow the deployment of
multiple cloudformation stacks. Previously, the stack was hard coded to
enable rapid development in the test environment. Now, the
infrastructure is ready to move to the next stage where the marketing
app can run on different stacks and accounts.

This refactoring includes:

1. Updating the config to specify VPC cidr blocks and AZs
2. Enabling region based VPC configs
3. Removing hard coded values and porting them to configs
4. Refactoring the deployer user to be 1:1 to the stack being deployed
5. Using Route 53 hosted zone delegation (`marketing.code.org`)

---

Route 53 Hosted Zone Delegation

Rather than having a flat domain structure (e.g.
`test-marketing.dev-code.org`, `prod-marketing.dev-code.org`) which
would require a mechanism to update the hosted zone for `dev-code.org`,
this PR proposes to use Route 53 hosted zone delegation. In this model,
`marketing.dev-code.org` is added as an `NS` entry and a hosted zone for
it is created. Each instance of this stack will then be permitted to add
subdomains to `marketing.dev-code.org`. This would mean that each stack
is equivalent to one subdomain of `marketing.dev-code.org`, for example,
`test.marketing.dev-code.org`, `prod.marketing.dev-code.org`.