refactor(TextField): use InputHTMLAttributes prop
this replaces the HTMLAttributes type that extends TextFieldProps with InputHTMLAttributes, which is a superset of HTMLAttributes. this gives us all possible input props without having to explicitly define all of them. props like placeholder, maxLength, disabled etc are all on InputHTMLAttributes but not HTMLAttributes even when HTMLInputElement is passed as the generic.

Unfortunately our 'size' prop on TextFieldProps conflicts with the native 'size' prop. rather than renaming our custom prop, I'm omitting it from the extended type