feat(marketing): add all-the-things snapshotting
This PR adds the ability to programmatically fork the "All The Things"
UI test page. When updating this page, a contributor can run the `fork`
command which will create a copy of the page in the `development` environment.
Contributors can then use this copy to test their changes. Once Contributors
wish to merge these changes with a higher environment, they can run the
`update-snapshot` command which will download the forked page and
overwrite the original page.

Finally, the `publish-snapshot` command will publish the forked page to
the authoritative copy of "All The Things" in the environment.

This new workflow allows for testing and validating changes to the "All The Things"
in a staged manner which prevents unexpected errors from being introduced
when deploying the marketing app to the various environments.
Additionally, changes to the page can now be reviewed as part of the
pull request process.

Longer term, the page should be automatically created from source control