fix(marketing): set page cache to 1 hour
This PR updates the default page caching behavior to cache pages for one hour.

An additional Cloudfront viewer response function was added because
Next.js has a bug[1] that appears to cause the `Cache-Control` header to
be missing when operating in a distributed environment. While waiting
for the bug to be fixed, the Cloudfront function will add the SWR cache
control header if the origin does not supply any cache control header.

If this workaround is not in place, Cloudfront will indefinetly cache
the page which is not what we intend for these pages.

[1] https://www.github.com/vercel/next.js/issues/73382