feat(marketing): increase task size and task count
This PR updates the number of tasks to a minimum of three to enable high
availability. Additionally, it balances these tasks across AZs to ensure
the tasks are spread out for high availability. Lastly, the task size is
updated to consume 1 vCPU (for the single threaded Next.js app) and 2 GB
of RAM.

Lastly, an autoscaling policy is added to allow scaling if the CPU >70%
as per the load test results.