fix(marketing): scroll to components before eyes
For firefox and webkit tests, eyes was previously running on the top of
the page but using a specific region. However, this results in some
screenshots being clipped due to image lazy loading. Rather, scroll to
the region before eyes screenshots.

This helps with the tracing as well to see what exactly the component
looks like so this is overall beneficial.