feat(newrelic): enable masked session replay and disable certain ajax endpoints
Add in support for 10% of erroring session replay to the New Relic
configuration. This will help aid in debugging sessions with errors to
see what may be occurring during a page load. The marketing site does
not render out any PII but we do have some input boxes. This PR masks
those inputs.

Also disable certain ajax endpoints which are not useful for health
monitoring.