fix(cloudfront): update cache policies
The marketing cdn currently does not set cache control
headers for the `/_next` behavior causing the dasiy chained distribution
to not cache these assets (always miss). Ensure that these assets are
immutably cached as they are hashed and are safe to be cached
indefinetly.

Additionally, disable caching on the main page paths as it causes a
double caching behavior which is quite confusing for content editors to
wait for content to become live.