feat(draft): auto-enable draft mode by hostname
To simplify content editing, this PR enables draft mode when using an
allowlisted draft domain rather than via the draft token route. This
allows content editors to share the direct address without any confusing
tokens in the same format - just with a different domain. The token was
originally added as a security measure against robotic traffic, but it
is not truly necessary at this time - and if robotic traffic does become
a problem we may pivot to using an idp which would be more controlled
anyway.